# deploy/couple-ordering-order-openapi.yaml
openapi: 3.0.3
info:
  title: Couple Ordering Order API
  version: 1.0.0
  description: Order APIs for the couple ordering mini program.
servers:
  - url: http://localhost:8070/api/v1
tags:
  - name: Order
    description: Order APIs
paths:
  /orders:
    post:
      tags:
        - Order
      summary: Submit an order from current shopping cart
      description: Creates an order from the current user's shopping cart, snapshots cart items into order_item, then clears the cart.
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            examples:
              normal:
                value:
                  remark: Less spicy
                  expectedTime: '2026-07-17 18:30:00'
      responses:
        '200':
          description: Created order detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResult'
        '409':
          description: Shopping cart is empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
    get:
      tags:
        - Order
      summary: Page current user's orders
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number, starts from 1.
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Page size.
      responses:
        '200':
          description: Paged order list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPageResult'
  /orders/{orderId}:
    get:
      tags:
        - Order
      summary: Get order detail
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/OrderIdPath'
      responses:
        '200':
          description: Order detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResult'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
  /orders/{orderId}/cancel:
    patch:
      tags:
        - Order
      summary: Cancel an order
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/OrderIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelOrderRequest'
            examples:
              normal:
                value:
                  cancelReason: Change of plans
      responses:
        '200':
          description: Canceled order detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResult'
        '409':
          description: Order cannot be canceled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
  /orders/{orderId}/complete:
    patch:
      tags:
        - Order
      summary: Complete an order
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/OrderIdPath'
      responses:
        '200':
          description: Completed order detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResult'
        '409':
          description: Order cannot be completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
components:
  parameters:
    UserIdHeader:
      name: X-User-Id
      in: header
      required: true
      schema:
        type: integer
        format: int64
        example: 1
      description: Temporary development user id header.
    OrderIdPath:
      name: orderId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
        example: 1
  schemas:
    CreateOrderRequest:
      type: object
      properties:
        remark:
          type: string
          maxLength: 255
          nullable: true
          example: Less spicy
        expectedTime:
          type: string
          nullable: true
          example: '2026-07-17 18:30:00'
    CancelOrderRequest:
      type: object
      properties:
        cancelReason:
          type: string
          maxLength: 255
          nullable: true
          example: Change of plans
    OrderItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        orderId:
          type: integer
          format: int64
          example: 1
        dishId:
          type: integer
          format: int64
          nullable: true
          example: 1
        dishName:
          type: string
          example: Tomato egg
        dishImageUrl:
          type: string
          nullable: true
          example: https://example.com/dish.jpg
        dishPrice:
          type: number
          format: double
          example: 20.00
        quantity:
          type: integer
          example: 2
        amount:
          type: number
          format: double
          example: 40.00
    OrderListItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        orderNumber:
          type: string
          example: '202607171830001231234'
        status:
          type: integer
          description: 1 pending, 2 accepted, 3 cooking, 4 completed, 5 canceled
          example: 1
        amount:
          type: number
          format: double
          example: 40.00
        totalQuantity:
          type: integer
          example: 2
        remark:
          type: string
          nullable: true
          example: Less spicy
        expectedTime:
          type: string
          nullable: true
          example: '2026-07-17T18:30:00'
        createTime:
          type: string
          example: '2026-07-17T18:00:00'
    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/OrderListItem'
        - type: object
          properties:
            userId:
              type: integer
              format: int64
              example: 1
            cancelReason:
              type: string
              nullable: true
              example: Change of plans
            updateTime:
              type: string
              example: '2026-07-17T18:00:00'
            items:
              type: array
              items:
                $ref: '#/components/schemas/OrderItem'
    OrderPage:
      type: object
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 10
        total:
          type: integer
          format: int64
          example: 23
        records:
          type: array
          items:
            $ref: '#/components/schemas/OrderListItem'
    OrderDetailResult:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/OrderDetail'
    OrderPageResult:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/OrderPage'
    ErrorResult:
      type: object
      properties:
        code:
          type: integer
          example: 40403
        message:
          type: string
          example: order not found
        data:
          nullable: true
          example: null